Update patch set 3
Patch Set 3:

There is no harm in not being able to load bt_vendor.conf as long as the correct configuration is set at compile time (via the BoardConfig option I mentioned earlier).

I also still insist that bluetoothd is not necessary and libbt-vendor should not be in proprietary files.

Patch-set: 3