Update patch set 2
Patch Set 2:

No, it's setting sys.cpu.mode. if you getprop sys.cpu.mode it will return 0, 1, or 2. These are defined in init.cardu.cpu.rc that is what is being set. So basically for it to set it boot it would just be setting the value of that property

Patch-set: 2