Update patch set 2
Patch Set 2:

It looks good, I verified it changes the proper value. Looks good with one exception. I think we need to have the back button on the icon in the upper left-hand corner to back out to the main settings fragment. I'm ready to push it as soon as that as fixed.

Patch-set: 2