Update patch set 1
Patch Set 1:

Only toggle support for now (no vibration). One side effect: passing the event to apps means that it also handle the event, and the app could duplicate the mute. Also, alarm app uses this events to dismiss the alarm, but the result is that the alarm is dismissed, but volume mute is toggle also (it think is better in this case ignore the key event in asusdec, but i don't find the way of distinguish what type of stream is sound, AudioManager#getMasterStreamType() always returns STREAM_MUSIC).

Patch-set: 1