Update patch set 1
Patch Set 1:

Actually, the asus key labels doesn't exists in the KeycodeLabels class and KeyLayoutMap fails to parse this keylayout file. This change mapped the unknown keycodes to F1..F7 keys (that are not supported by tf700t keyboard). The new F1..F7 keycodes are all handled by the asusdec class KeyHandler and all of them are consumed by this class, so they never be passed to apps (before and after this patch). Is only an aesthetical change that allow to load the rest of the keys, neither of them has any special function, except BACK and POWER (which is not used in asusdec; it's handled by gpio-keys.kl)

Patch-set: 1