Update patch set 4
Patch Set 4:

With a clean build i have the same crashes.

BTW, the first time i try to set "setenforce 1" the system get collapsed and i have to reboot the device. The second time it allows it, but start crashing when an app access the storage volumen.

"getenforce" is denied for root user

Patch-set: 4